requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.1.0
babel==2.16.0
certifi==2023.11.17
blinker==1.7.0
selenium-wire==5.1.0
cffi==1.16.0
charset-normalizer==3.3.2
exceptiongroup==1.2.0
execnet==2.0.2
h11==0.14.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
minio==7.2.4
outcome==1.3.0.post0
packaging==23.2
pika==1.3.2
pluggy==1.3.0
py==1.11.0
pycparser==2.21
pycryptodome==3.20.0
PySocks==1.7.1
pdfplumber==0.11.4
pytest==7.4.3
pytest-base-url==2.0.0
pytest-rerunfailures==14.0
pytest-forked==1.6.0
pytest-html==4.1.1
pytest-metadata==3.0.0
pytest-selenium==4.0.2
pytest-variables==3.0.0
pytest-xdist==2.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-interface==1.6.1
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tenacity==8.2.3
toml==0.10.2
tomli==2.0.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==1.26.18
webdriver-manager==4.0.1
wsproto==1.2.0
xmltodict==0.13.0
